requirements.txt
alembic==1.13.2
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.3
numpy==2.0.0
packaging==24.1
pandas==2.2.2
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.0
scipy==1.14.0
six==1.16.0
SQLAlchemy==2.0.31
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
gunicorn
marshmallow-sqlalchemy
